+++ Nickel.InfoScreens/InfoScreenReplacementRoute.cs
@@ -1,24 +1,31 @@
-﻿namespace Nickel.InfoScreens;
+﻿using Newtonsoft.Json;
 
-internal sealed class InfoScreenReplacementRoute(InfoScreenEntry entry, Route route, Route? originalRoute) : Route
+namespace Nickel.InfoScreens;
+
+internal sealed class InfoScreenReplacementRoute : Route
 {
-	public readonly InfoScreenEntry Entry = entry;
+	[JsonIgnore]
+	public required InfoScreenEntry? Entry { get; init; }
+	
+	public required Route Route { get; init; }
+	public required Route? OriginalRoute { get; init; }
 	
 	public override bool TryCloseSubRoute(G g, Route r, object? arg)
 	{
-		if (route.TryCloseSubRoute(g, r, arg))
+		if (this.Route.TryCloseSubRoute(g, r, arg))
 			return true;
-		if (r != route)
+		if (r != this.Route)
 			return false;
 		
-		ModEntry.Instance.OnClose(g, this.Entry);
+		if (this.Entry is not null)
+			ModEntry.Instance.OnClose(g, this.Entry);
 		if (g.state.route is Combat combat)
-			combat.routeOverride = originalRoute;
-		else if (originalRoute is not null)
-			g.state.route = originalRoute;
+			combat.routeOverride = this.OriginalRoute;
+		else if (this.OriginalRoute is not null)
+			g.state.route = this.OriginalRoute;
 		return true;
 	}
 
 	public override void Render(G g)
-		=> route.Render(g);
+		=> this.Route.Render(g);
 }